+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Features
 - Added `Template` model and created pivot tables for Templates (120757529)
+- Added additional flow for sending `Email Applications` (120758677)
+- Added `Templates Controller` and ability to send emails stored in the database (120765999)
 
 # 1.4.0
 2016-06-20